Perhaps rectangles with dimensions and position are better

+++ Engine/Vec2.cpp
@@ -1,6 +1,6 @@
 #include "Vec2.h"
 
-Vec2::Vec2(float x, float y)
+Vec2::Vec2(int x, int y)
 {
 	this->x = x;
 	this->y = y;
