+++ Engine/Vec2.h
@@ -3,9 +3,9 @@
 class Vec2
 {
 public:
-	float x;
-	float y;
+	int x;
+	int y;
 public:
 	Vec2() = default;
-	Vec2(float x, float y);
+	Vec2(int x, int y);
 };
